unifying the two look-for-applicable-parent-strategy files
